refactored player details to it's own domain component
