--- src/components/PlayerInfo/PlayerInfo.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function PlayerInfo({ player }) {
-  return <div>Not sure if i'll use this</div>;
-}
-
-export default PlayerInfo;
