--- src/components/PlayerInfo/index.js
@@ -1,2 +1,0 @@
-export * from "./PlayerInfo";
-export { default } from "./PlayerInfo";
