+++ src/domain/PlayerDetails/PlayerDetails.module.css
@@ -1,0 +1,3 @@
+.temp {
+  
+}
