+++ src/domain/PlayerDetails/index.js
@@ -1,0 +1,2 @@
+export * from "./PlayerDetails";
+export { default } from "./PlayerDetails";
